Backend/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose
python-multipart
email-validator
